Moves to separate modules complicated opcodes
