+++ csound-expression-typed-core/src/Csound/Typed/Core/Types/SE/Ref.hs
@@ -4,7 +4,6 @@ module Csound.Typed.Core.Types.SE.Ref
   ( Ref (..)
   , newRef
   , newLocalRef
-  , newGlobalRef
   ) where
 
 import Control.Monad
@@ -19,6 +18,7 @@ import Csound.Typed.Core.Types.Tuple
 -- | It describes a reference to mutable values.
 newtype Ref a = Ref [Var]
 
+{-
 -- | It creates global variables if it is used from top level instrument
 -- or global variable if it's used inside local instrument
 --
@@ -31,11 +31,12 @@ newRef initVals = do
   if isGlobal
     then newGlobalRef initVals
     else newLocalRef initVals
+-}
 
 -- | Creates global mutable variable (reference). It can be shared between different
 -- instances ofthe instruments.
-newGlobalRef :: forall a . Tuple a => a -> SE (Ref a)
-newGlobalRef initVals =
+newRef :: forall a . Tuple a => a -> SE (Ref a)
+newRef initVals =
   fmap Ref $ SE $ lift $ zipWithM State.initGlobalVar (tupleRates @a) =<< fromTuple initVals
 
 -- | Creates local mutable variable (reference). It can not be shared between different local instruments
